日更

+++ TiHMS/CommonModule/src/main/java/tiquartet/CommonModule/vo/CreditVO.java
@@ -2,8 +2,8 @@ package tiquartet.CommonModule.vo;
 
 public class CreditVO {
     public int changeType;
-    public int change;
-    public int balance;
+    public double change;
+    public double balance;
     public long orderID;
     public long creditRecordID;
 }
